requirements.txt
flask
flask_cors
Flask-MongoAlchemy
Flask-PyMongo
auth0-python
nose
pymongo
bcrypt